HDOP sent as float
fixed issue https://github.com/jcaridadhdez/RAK811-tracker/issues/3

+++ lib/system_utils/gps.h
@@ -136,7 +136,7 @@ int16_t GpsGetLatestGpsAltitude( void );
  *
  * \retval hdop
  */
-int8_t GpsGetLatestGpsHorizontalDilution ( void );
+float GpsGetLatestGpsHorizontalDilution ( void );
 
 
 /*!
